stock/company-deep-research/pyproject.toml
[project]
name = "company-deep-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "langchain>=1.2.12",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.3",
    "langsmith>=0.7.20",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "streamlit>=1.55.0",
    "tavily-python>=0.7.23",
    "uvicorn>=0.42.0",
    "yfinance>=1.2.0",
]
